% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulkImport.R
\name{bulkImport}
\alias{bulkImport}
\title{The Need for Speed: Bulk Import}
\usage{
bulkImport(
  FileExtension,
  fread = FALSE,
  FolderPath = "./Input",
  ToDF = FALSE,
  xlsheet = 1
)
}
\arguments{
\item{FileExtension}{Supported format from list of currently supported formats. Also used to remove file extension from element name.}

\item{fread}{(\code{TRUE}/\code{FALSE}) If \code{TRUE}, will use fread to read as DT. Although it is faster, this option is only available for csv format only. Default is \code{FALSE}.}

\item{FolderPath}{Where target files are located. Default is \code{"./Input"}.}

\item{ToDF}{(\code{TRUE}/\code{FALSE}) If \code{TRUE}, will convert back to class \code{data.frame}. IF \code{FALSE}, will remain class \code{data.table}.}

\item{xlsheet}{Which Excel sheet to read and import. Default is \code{"1"}.}
}
\value{
List of DF or DT where each element of the list (same name as file name) contains the contents of a single imported file.
}
\description{
Mass import files from specified Path + File Extension. All cols imported as class char to avoid most formatting issues.
}
\details{
Requires the following packages in addition to base R packages: \code{tidyverse}, \code{data.table}, \code{read_xl}.

Currently supported formats are: \code{.csv}, \code{.xls}, \code{.xlsx}

Function will stop and output error message if FileExtension if not from list of supported formats.
}
