% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTfilter.R
\name{DTfilter}
\alias{DTfilter}
\title{The Need for Speed: Fast Subset}
\usage{
DTfilter(dfName, purpose, varName, toDF = FALSE)
}
\arguments{
\item{dfName}{Input DF.}

\item{purpose}{(\code{"keep"}/\code{"remove"}) Use \code{"keep"} to keep and \code{"remove"} to remove the specified cols.}

\item{varName}{Names of vars. Can be in form of : col num, string, or list of strings.}

\item{toDF}{(\code{TRUE}/\code{FALSE}) If \code{TRUE}, will convert back to class DF Default is \code{FALSE}.}
}
\value{
Either DF or DT & keep or remove specified cols.
}
\description{
Fast subset of DF cols using DT tools.
}
