% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_Excel_date.R
\name{fix_Excel_date}
\alias{fix_Excel_date}
\title{Fix Excel date format issue}
\usage{
fix_Excel_date(value)
}
\arguments{
\item{value}{Excel date entry.}
}
\value{
Tidy date format in form of \verb{\%m/\%d/\%y}. Can further modify using Lubridate package.
}
\description{
Converts Excel date to a format R/Lubridate can work with.
}
