% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_n_fill.R
\name{match_n_fill}
\alias{match_n_fill}
\title{Fill in target DF using matches from another DF}
\usage{
match_n_fill(ref_1, ref_2, fill_with)
}
\arguments{
\item{ref_1}{Matching var/entry in source DF.}

\item{ref_2}{Matching var/entry in target DF.}

\item{fill_with}{Target DF.}
}
\value{
Target DF with additional content.
}
\description{
Very common occurrence. Very useful. 10/10, would recommend.
}
