% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/origLvl.R
\name{origLvl}
\alias{origLvl}
\title{Show/Count distinct levels}
\usage{
origLvl(dfName, varName = NULL, length = FALSE)
}
\arguments{
\item{dfName}{Input DF. Can use \code{$} notation for both single DF or nested DF.}

\item{varName}{Can be used to pass either var name or col num. Default is NULL.}

\item{length}{(TRUE/FALSE) If FALSE, will print out all levels. If TRUE, will only print out length. Default is FALSE.}
}
\value{
Prints either all unique levels within specified DF$var or just the length in console.
}
\description{
A custom function/wrapper for unique() and length(unique()). Use-case depends on user. Can use \code{$} or \verb{[[]]} notation.
}
