% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr_text.R
\name{qr_text}
\alias{qr_text}
\title{Extract Text from QR Code}
\usage{
qr_text(imagePath, loadLibrary = FALSE)
}
\arguments{
\item{imagePath}{Path of image to extract text from.}

\item{loadLibrary}{Will load required libraries (quadrangle & ggplot2) if set to \code{TRUE}. Default is \code{FALSE}.}
}
\value{
Extracted text in console.
}
\description{
Uses \href{https://github.com/brianwdavis/quadrangle}{Quadrangle} to read and extract text from qr code.
}
