% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmAllNA.R
\name{rmAllNA}
\alias{rmAllNA}
\title{Out of sight, out of mind}
\usage{
rmAllNA(dfName)
}
\arguments{
\item{dfName}{Input DF.}
}
\value{
DF where columns with 100\% NA are removed.
}
\description{
Why waste storage saving a bunch of unnecessary all NA columns when you can remove it to save on computation time and costs? This function removes all cols where everything is NA.
}
\details{
User needs to manually save results.
}
