% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmSomeNA_col.R
\name{rmSomeNA_col}
\alias{rmSomeNA_col}
\title{Out of sight, out of mind (col version)}
\usage{
rmSomeNA_col(dfName, comparisonSign, threshold)
}
\arguments{
\item{dfName}{Input DF.}

\item{comparisonSign}{Options are ">" and ">=".}

\item{threshold}{Col missingness threshold.}
}
\value{
DF where columns with threshold NA are removed.
}
\description{
Removes col(s) that are equal to/exceeds col missingness threshold.
}
\examples{
#All columns with > 0.1 (10\% missingness) will be removed.
rmSomeNA_col(df,">",0.1)
}
