% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmSomeNA_row.R
\name{rmSomeNA_row}
\alias{rmSomeNA_row}
\title{Out of sight, out of mind (row version)}
\usage{
rmSomeNA_row(dfName, comparisonSign, threshold)
}
\arguments{
\item{dfName}{Input DF.}

\item{comparisonSign}{Options are ">" and ">=".}

\item{threshold}{Row missingness threshold.}
}
\value{
DF where rows with threshold NA are removed.
}
\description{
Removes row(s) that are equal to/exceeds row missingness threshold.
}
\examples{
#All rows with > 0.1 (10\% missingness) will be removed.
rmSomeNA_row(df,">",0.1)
}
