% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleLvls.R
\name{singleLvls}
\alias{singleLvls}
\title{Levels within specified DF$var (Depreciated)}
\usage{
singleLvls(dfName, varName, showLevels = TRUE)
}
\arguments{
\item{dfName}{Input DF.}

\item{varName}{Input var.}

\item{showLevels}{(\code{TRUE}/\code{FALSE}) If \code{TRUE}, will print out all levels. If \code{FALSE}, will only print out length. Default is \code{TRUE}.}
}
\value{
Prints either all unique levels within specified DF$var or just the length in console.
}
\description{
Show/Count number of unique levels within specified DF$var.
}
\details{
For DF nested within a list, use \code{$} notation.
}
