% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soleImport.R
\name{soleImport}
\alias{soleImport}
\title{The Need for Speed: Sole Import}
\usage{
soleImport(File, fread = FALSE, ToDF = FALSE, xlsheet = 1)
}
\arguments{
\item{File}{File Path + Name + Extension of file.}

\item{fread}{(\code{TRUE}/\code{FALSE}) If \code{TRUE}, will use fread to read as DT. Although it is faster, this option is only available for csv format only. Default is \code{FALSE}. If \code{TRUE}, will replace all "" with \code{NA}.}

\item{ToDF}{(\code{TRUE}/\code{FALSE}) If \code{TRUE}, will convert back to class \code{data.frame}. IF \code{FALSE}, will remain class \code{data.table}.}

\item{xlsheet}{Which Excel sheet to read and import. Default is \code{"1"}.}
}
\value{
Single DF/DT containing contents of imported file.
}
\description{
Import specified file with all cols imported as class char to avoid most formatting issues.
}
\details{
Requires the following packages in addition to base R packages: \code{tidyverse}, \code{data.table}, \code{read_xl}.

Currently supported formats are: \code{.csv}, \code{.xls}, \code{.xlsx}
}
