% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vitae_data.R
\name{vitae_data}
\alias{vitae_data}
\title{vitae_data: Vitae Data Template}
\usage{
vitae_data()
}
\value{
Dataframe template for user to export. The user can decide which package to use for exporting to \code{.xlsx} format.
}
\description{
Creates dataframe to use in tandem with \verb{Vitae Rmd} Rmarkdown template to semi-automate the resme and/or CV creation process.
}
